Address analyzer complaints, move packaging to separate build job (#797)
* move packaging to a separate job, only use Deterministic and SourceLink when building packages
* Fixes for IDE0003, IDE0007, IDE0019, IDE0090, IDE0038, S1905, S1939
* use switch expressions, conditionals, simplified null checks
* ignore some rules
* drop codecov

+++ src/CircuitBreaker/test/HystrixBase.Test/Metric/CommandStreamTest.cs
@@ -24,7 +24,7 @@ namespace Steeltoe.CircuitBreaker.Hystrix.Metric.Test
             base.Dispose();
         }
 
-        private static readonly AtomicInteger UniqueId = new AtomicInteger(0);
+        private static readonly AtomicInteger UniqueId = new (0);
 
         public class Command : HystrixCommand<int>
         {
