+++ src/CircuitBreaker/test/HystrixBase.Test/Util/MockedTime.cs
@@ -8,7 +8,7 @@ namespace Steeltoe.CircuitBreaker.Hystrix.Util.Test
 {
     internal class MockedTime : ITime
     {
-        private readonly AtomicInteger time = new AtomicInteger(0);
+        private readonly AtomicInteger time = new (0);
 
         public long CurrentTimeInMillis
         {
